plancheck is now a subcommand instead of a flag
